feat: App 컴포넌트, 홈페이지 분리
